+++ source/views/ScanPane.svelte
@@ -13,20 +13,13 @@
   import debug from "../lib/debug.ts";
   import Navbar from "./components/nav/Navbar.svelte";
   import type { Tab } from "./components/nav/tab.ts";
-  import ListTitle from "./components/ListTitle.svelte";
-  import CssTarget from "./components/CssTarget.svelte";
-  import DeleteButton from "./components/buttons/DeleteButton.svelte";
+  import DomSelect from "./components/DomSelect.svelte";
 
   let NavTabs: Tab[] = $state([
     { label: "plugins", title: "Plugin Selection" },
     { label: "dom-selection", title: "DOM Selection" },
   ]);
   let currentTab: string = $state("plugins");
-
-  const Selections = {
-    SpecifyTarget: "specifytarget",
-    FullScan: "fullscan",
-  };
 
   let statusMessage: string | null = $state(null);
   let results: Results = $state([]);
@@ -96,25 +89,6 @@
       if (e instanceof Error) statusMessage = e.message;
     }
   }
-
-  let selection: string = $state(Selections.SpecifyTarget);
-
-  let domSelectComponents: { id: number }[] = $state([]);
-  let domSelectComponentCount: number = 0;
-
-  function handleAddButtonClick(): void {
-    debug.debug("New dom selector element button was clicked");
-    domSelectComponents = [
-      ...domSelectComponents,
-      { id: domSelectComponentCount++ },
-    ];
-  }
-
-  function removeDomComponent(id: number): void {
-    domSelectComponents = domSelectComponents.filter(
-      (component) => component.id !== id,
-    );
-  }
 </script>
 
 <Navbar bind:Tabs={NavTabs} bind:current={currentTab} />
@@ -127,41 +101,7 @@
 
     <!--DOM Selection and entire website scan-->
   {:else if currentTab == "dom-selection"}
-    <div class="radio-choice">
-      <label>
-        <input
-          type="radio"
-          name="Specify Target"
-          value={Selections.SpecifyTarget}
-          bind:group={selection}
-        />
-        Specify Target
-      </label>
-      <label>
-        <input
-          type="radio"
-          name="Full Scan"
-          value={Selections.FullScan}
-          bind:group={selection}
-        />
-        Full Scan
-      </label>
-    </div>
-
-    {#if selection === Selections.SpecifyTarget}
-      <ListTitle title="Targets" onAdd={handleAddButtonClick} />
-      {#each domSelectComponents as component (component.id)}
-        <CssTarget></CssTarget>
-        <DeleteButton
-          onDelete={() => {
-            removeDomComponent(component.id);
-          }}
-        />
-        <hr />
-      {/each}
-    {:else if selection === Selections.FullScan}
-      <h3>Noting to Pick 😄👍</h3>
-    {/if}
+    <DomSelect></DomSelect>
   {/if}
 
   <button onclick={startScan}>Scan Now</button>
@@ -198,10 +138,5 @@
   .container {
     margin-top: 15px;
     margin-bottom: 15px;
-  }
-  .radio-choice {
-    display: flex;
-    gap: 20px;
-    margin-bottom: 10px;
   }
 </style>
